requirements.txt
numpy
pandas
matplotlib
scikit-learn
openai==0.28.0
tqdm
sentence-transformers
torch
umap-learn
lexrank